Create RaceViewer
